Move convenient query methods to Providers. Change Data provider implementation to map-based to avoid concurrency issues.
