+++ app/src/main/java/com/example/aikataulu/database/TimetableDbHelper.kt
@@ -31,60 +31,4 @@ class TimetableDbHelper(context: Context) :
     override fun onDowngrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
         onUpgrade(db, oldVersion, newVersion)
     }
-
-    fun getStopByHrtId(hrtId: String): Cursor? {
-        val db = readableDatabase
-        val entry = StopContract.StopEntry
-        val cursor = db.query(
-            entry.TABLE_NAME,
-            entry.allColumns(),
-            "${entry.COLUMN_NAME_HRTID} = ?",
-            arrayOf(hrtId),
-            null,
-            null,
-            null
-        )
-        return cursor
-    }
-
-    fun getConfigByWidgetId(widgetId: Int): Cursor? {
-        val db = readableDatabase
-        val entry = ConfigurationContract.ConfigurationEntry
-        val cursor = db.query(
-            entry.TABLE_NAME,
-            entry.allColumns(),
-            "${entry.COLUMN_NAME_WIDGET_ID} = ?",
-            arrayOf(widgetId.toString()),
-            null,
-            null,
-            null
-        )
-        return cursor
-    }
-
-    fun updateInterval(widgetId: Int, newInterval: Int) {
-        /*
-        Proper way:
-        val db = dbHelper.writableDatabase
-
-        // New value for one column
-        val title = "MyNewTitle"
-        val values = ContentValues().apply {
-            put(FeedEntry.COLUMN_NAME_TITLE, title)
-        }
-
-        // Which row to update, based on the title
-        val selection = "${FeedEntry.COLUMN_NAME_TITLE} LIKE ?"
-        val selectionArgs = arrayOf("MyOldTitle")
-        val count = db.update(
-                FeedEntry.TABLE_NAME,
-                values,
-                selection,
-                selectionArgs)
-
-         */
-        /* Easy way: */
-        val entry = ConfigurationContract.ConfigurationEntry
-        writableDatabase.execSQL("UPDATE ${entry.TABLE_NAME} SET ${entry.COLUMN_NAME_UPDATE_INTERVAL_SECONDS} = $newInterval WHERE ${entry.COLUMN_NAME_WIDGET_ID} = $widgetId")
-    }
 }
